src/python/requirements.txt
argcomplete>=1.0.0
attrdict>=2.0.0
certifi>=2015.11.20
colorclass>=2.2.0
colorlog>=2.6.0
coverage>=4.0.3
enum34>=1.1.1
faulthandler>=2.4
filelock>=2.0.5
future>=0.15.2
futures>=3.0.3
logilab-common>=1.1.0
nose>=1.3.7
paramiko>=1.16.0
pytz>=2015.7
PyYAML>=3.12
redis>=2.10.5
remember>=0.1
requests>=2.12.3
requests_cache>=0.4.10
setproctitle>=1.1.9
sh>=1.11
six>=1.10.0
subprocess32>=3.2.7
termcolor>=1.1.0
terminaltables>=3.0.0
tzlocal>=1.2
